Создан модуль basketModule в store
