+++ test_project/src/store/index.js
@@ -2,12 +2,13 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import createPersistedState from 'vuex-persistedstate';
 import {productsModule} from "./modules/productsModule";
+import {basketModule} from "./modules/basketModule";
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   state: {
-    basket:[],
+    // basket:[],
     activeLink:'Электроприборы',
     activeCategoryMebel:'Стулья',
     activeCategoryElectro:'Светильники',
@@ -16,19 +17,12 @@ export default new Vuex.Store({
     description:[]  //массив описаний товаров по id товара
   },
   getters:{
-      //получить содержимое корзны
-    getCart(state){
-        return state.basket//эта функция возвращает корзину из хранилища
-    },
-      //получить информацию о продукте
+    //получить информацию о продукте
     getInfo(state){
         return state.infoAboutProduct
     },
     getShowloader(state){
         return state.showloader
-    },
-    getBasket(state){
-        return state.basket
     },
     getActiveLink(state){
         return state.activeLink
@@ -44,24 +38,6 @@ export default new Vuex.Store({
     }
   },
   mutations: {
-    setBasket(state,payload){ state.basket = [...state.basket, ...payload]},
-      //добавить в корзину
-    addToCart(state, product) {
-        const element = Object.assign(product);
-        state.basket.push(element);
-        console.log("state.basket=", state.basket)
-    },
-      //Удалить из корзины
-    deleteProduct(state, product){
-        let find = state.basket.find((element) => element.id === product.id);//находим необходимый продукт
-        //удаляем из корзины товаров 1 товар с порядкового номера това с выбранным id
-        this.state.basket.splice(state.basket.indexOf(find),1)//удаляем из массива корзины 1 товар начиная с индекса выбранного товара
-        console.log(state.basket);
-    },
-      //очистить корзину
-    clearBasket(state){
-        state.basket = []
-    },
       //установить в стейт информацию о продукте
     setInfoAboutProduct(state, payload){
         state.infoAboutProduct = [...state.infoAboutProduct, ...payload]
@@ -88,19 +64,6 @@ export default new Vuex.Store({
     }
   },
   actions: {
-      //загрузить корзину
-    loadToCart({commit}, product) {
-          commit('addToCart', product)//передаем в коммит good.id, чтобы вызвать мутацию addToCart
-      },
-      //get запрос при заказе товара с удачным результатом
-    loadOrder({commit}){
-        return fetch('http://test1.web-gu.ru/?action=send_form')
-            .then(response => response.json())
-            .then(data => {
-                console.log('data=', data);
-                commit('clearBasket')
-            });
-    },
     //загрузить описание к товарам
       loadDescription({commit}){
           return fetch('http://test1.web-gu.ru')
@@ -136,7 +99,8 @@ export default new Vuex.Store({
   },
   //плагин для сохранения состояния конкретных ключей (у нас нет бэка, поэтому нужен плагин)
   modules:{
-      productsModule
+      productsModule,
+      basketModule
   },
   plugins: [
         createPersistedState({
